use consistent type for window num
all the functions in bios/mono.c use short, so still somewhat broken

+++ source/bios/mono.h
@@ -94,7 +94,7 @@ extern void msetcursor( int row, int col );
 
 //==========================================================================
 // Refreshes a window
-void mrefresh(short n);
+void mrefresh(int n);
 
 #else			//null versions for when debugging turned off
